+++ CHANGES.md
@@ -334,3 +334,7 @@
 - Protect links in script to prevent take-down
 - Minor improve error handling if unable to get proper file data stream
 - Minor display adjustment for version menu
+
+### Pre-release (work-in-progress)
+- Oticon: move legacy versions checker into current Genie 2 checker (please edit `config.ini` referring to [Versions](https://github.com/JediLin/Hearing-Aids-Fitting-Software-Update-Checkers/wiki/Versions) information from our Wiki if need to check previous versions)
+- Remove unused scripts